+++ perl/FSM/HDL/FlattenedDT.pm
@@ -14,6 +14,7 @@ use FSM::ExpressionNamer;
 use FSM::Synthesis::EnableGraph;
 use FSM::Synthesis::EnableGraph::FactorizationSupport;
 use FSM::Synthesis::EnableGraph::IntermediateSignalSupport;
+use FSM::Synthesis::EnableGraph::ModulePlanningSupport;
 use FSM::HDL::FlattenedDT::Orchestrator;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ASTFactorizationSupport;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ConsolidatedIntermediateEmitter;
@@ -85,6 +86,7 @@ sub new ($class, %args) {
     $self->{enable_graph} = FSM::Synthesis::EnableGraph->new(flattened_dt => $self);
     $self->{enable_graph_factorization_support} = FSM::Synthesis::EnableGraph::FactorizationSupport->new(flattened_dt => $self);
     $self->{enable_graph_intermediate_support} = FSM::Synthesis::EnableGraph::IntermediateSignalSupport->new(flattened_dt => $self);
+    $self->{enable_graph_module_planning_support} = FSM::Synthesis::EnableGraph::ModulePlanningSupport->new(flattened_dt => $self);
     $self->{orchestrator} = FSM::HDL::FlattenedDT::Orchestrator->new(flattened_dt => $self);
     $self->{backend_sv_scaffold} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::ScaffoldEmitter->new(flattened_dt => $self);
     $self->{backend_sv_internal_decl} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::InternalDeclarationEmitter->new(flattened_dt => $self);